CORE-7645 Moving handlers from AppCategory to OntologyHierarchy
